requirements.txt
torch
transformers
accelerate
tqdm
pytest
Pillow
mtcnn
tensorflow
scikit-learn
